added loadList, loadDetails, ShowLoadingMessage, and hideLoadingMessage functions
loadList function gets the pokemon info from the pokeapi and fills out the pokemonList array.
loadDetails function takes the object from the array for the button clicked and logs the pokemons name, picture , height, and types to the console.
ShowloadingMessage function add a loading message to the page while loadList and loadDetails in retrieving the data from pokeapi.
hideLoadingMessage function removed the loading message once the data has been retrieved.